- Added the CloudyWebAPI module - Took out all networking code from the CloudySaveManager module - Added all networking code to the CloudyWebAPI module - Updated README.md with instructions on using CloudyWebAPI.
