+++ Source/CloudySaveManager/Private/CloudySaveManager.cpp
@@ -1,41 +1,27 @@
 #include "CloudySaveManagerPrivatePCH.h"
  
 #include "CloudySaveManager.h"
+#include "../../CloudyWebAPI/Public/ICloudyWebAPI.h"
+
 #include "PlatformFeatures.h"
 #include "GameFramework/SaveGame.h"
-#include "HttpRequestAdapter.h"
-#include "HttpModule.h"
-#include "IHttpResponse.h"
-#include "Base64.h"
-#include "AllowWindowsPlatformTypes.h"
-#include "ThirdParty/libcurl/include/Windows/curl/curl.h"
-#include "HideWindowsPlatformTypes.h"
-#include "string"
+
+DEFINE_LOG_CATEGORY(CloudySaveManagerLog)
 
 static const int UE4_SAVEGAME_FILE_TYPE_TAG = 0x53415647;		// "sAvG"
 static const int UE4_SAVEGAME_FILE_VERSION = 1;
-static const FString BaseUrl = "http://127.0.0.1:8000";
-static const FString AuthUrl = "/api-token-auth/";
-static const FString SaveDataUrl = "/save-data/";
-static FString Token;
 
 // Automatically starts and shuts down when UE4 is started/closed
 void CloudySaveManagerImpl::StartupModule()
 {
-    UE_LOG(LogTemp, Warning, TEXT("CloudySaveManager started"));   
+    UE_LOG(CloudySaveManagerLog, Warning, TEXT("CloudySaveManager started"));
 }
  
 void CloudySaveManagerImpl::ShutdownModule()
 {
-    UE_LOG(LogTemp, Warning, TEXT("CloudySaveManager stopped"));
+    UE_LOG(CloudySaveManagerLog, Warning, TEXT("CloudySaveManager stopped"));
 }
 
-static size_t WriteCallback(void *contents, size_t size, size_t nmemb, void *userp)
-{
-    ((std::string*)userp)->append((char*)contents, size * nmemb);
-    return size * nmemb;
-}
- 
 bool CloudySaveManagerImpl::Cloudy_SaveGameToSlot(USaveGame* SaveGameObject, const FString& SlotName, 
                                                   const int32 UserIndex, const int32 PCID)
 {
@@ -75,155 +61,10 @@ bool CloudySaveManagerImpl::Cloudy_SaveGameToSlot(USaveGame* SaveGameObject, con
         // Stuff that data into the save system with the desired file name
         SaveSystem->SaveGame(false, *SlotName, UserIndex, ObjectBytes);
 
-        // Token variable will be populated with robot's token
-        AttemptAuthentication(TEXT("joel"), TEXT("1234"));
-        bSuccess = true; 
+        bSuccess = ICloudyWebAPI::Get().UploadFile(SlotName, PCID);
     }
 
     return bSuccess;
-}
-
-bool CloudySaveManagerImpl::AttemptAuthentication(FString Username, FString Password)
-{
-    bool RequestSuccess = false;
-
-    FString Url = BaseUrl + AuthUrl; // "http://127.0.0.1:8000/api-token-auth/";
-    FString ContentString;
-
-    TSharedPtr<FJsonObject> JsonObject = MakeShareable(new FJsonObject());
-    JsonObject->SetStringField(TEXT("username"), Username);
-    JsonObject->SetStringField(TEXT("password"), Password);
-
-    TSharedRef<TJsonWriter<TCHAR>> JsonWriter = TJsonWriterFactory<>::Create(&ContentString);
-    FJsonSerializer::Serialize(JsonObject.ToSharedRef(), JsonWriter);
-
-    TSharedRef<IHttpRequest> HttpRequest = FHttpModule::Get().CreateRequest();
-    HttpRequest->SetHeader(TEXT("Content-Type"), TEXT("application/json"));
-    HttpRequest->SetURL(Url);
-    HttpRequest->SetVerb(TEXT("POST"));
-    HttpRequest->SetContentAsString(ContentString);
-    HttpRequest->OnProcessRequestComplete().BindRaw(this, &CloudySaveManagerImpl::OnAuthResponseComplete);
-    RequestSuccess = HttpRequest->ProcessRequest();
-
-    UE_LOG(LogTemp, Warning, TEXT("URL = %s"), *Url);
-    UE_LOG(LogTemp, Warning, TEXT("ContentString = %s"), *ContentString);
-
-    return RequestSuccess;
-}
-
-bool CloudySaveManagerImpl::UploadFile(FString Filename, int32 PlayerControllerId)
-{
-    bool RequestSuccess = false;
-
-    CURL *curl;
-    CURLcode res;
-    std::string readBuffer;
-    
-    FString Url = BaseUrl + SaveDataUrl;
-    std::string UrlCString(TCHAR_TO_UTF8(*Url));
-    
-    // Filepath of .sav file
-    FString Filepath = FPaths::GameDir();
-    Filepath += "Saved/SaveGames/" + Filename + ".sav";
-    std::string filePath(TCHAR_TO_UTF8(*Filepath));
-    
-    // Get game name
-    FString GameName = FApp::GetGameName();
-    std::string gameName(TCHAR_TO_UTF8(*GameName));
-    
-    // Convert PlayerControllerId
-    FString playerControllerIdFString = FString::FromInt(PlayerControllerId);
-    std::string playerControllerId(TCHAR_TO_UTF8(*playerControllerIdFString));
-    
-    struct curl_httppost *formpost = NULL;
-    struct curl_httppost *lastptr = NULL;
-    struct curl_slist *headerlist = NULL;
-    //static const char buf[] = "Expect:";
-    FString AuthHeader = "Authorization: Token " + Token;
-    std::string AuthHeaderCString(TCHAR_TO_UTF8(*AuthHeader));
-    
-    curl_global_init(CURL_GLOBAL_ALL);
-    
-    /* Fill in the file upload field */
-    curl_formadd(&formpost, &lastptr,
-        CURLFORM_COPYNAME, "saved_file",
-        CURLFORM_FILE, filePath.c_str(),
-        CURLFORM_END);
-    
-    /* Fill in the player controller ID field */
-    curl_formadd(&formpost, &lastptr,
-        CURLFORM_COPYNAME, "controller",
-        CURLFORM_COPYCONTENTS, playerControllerId.c_str(),
-        CURLFORM_END);
-    
-    /* Fill in the game name field */
-    curl_formadd(&formpost, &lastptr,
-        CURLFORM_COPYNAME, "game",
-        CURLFORM_COPYCONTENTS, gameName.c_str(),
-        CURLFORM_END);
-    
-    curl = curl_easy_init();
-    /* initialize custom header list (stating that Expect: 100-continue is not
-    wanted */
-    headerlist = curl_slist_append(headerlist, AuthHeaderCString.c_str());
-    if (curl) {
-        /* what URL that receives this POST */
-        curl_easy_setopt(curl, CURLOPT_URL, UrlCString.c_str());
-
-        /* only disable 100-continue header if explicitly requested */
-        curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headerlist);
-
-        /* What form to send */
-        curl_easy_setopt(curl, CURLOPT_HTTPPOST, formpost);
-    
-        /* Set up string to write response into */
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &readBuffer);
-    
-        /* Perform the request, res will get the return code */
-        res = curl_easy_perform(curl);
-    
-        /* always cleanup */
-        curl_easy_cleanup(curl);
-    
-        /* then cleanup the formpost chain */
-        curl_formfree(formpost);
-        /* free slist */
-        curl_slist_free_all(headerlist);
-    
-        UE_LOG(LogTemp, Warning, TEXT("Response data: %s"), UTF8_TO_TCHAR(readBuffer.c_str()));
-    }
-
-    return RequestSuccess;
-}
-
-void CloudySaveManagerImpl::OnAuthResponseComplete(FHttpRequestPtr Request, 
-                                                   FHttpResponsePtr Response, bool bWasSuccessful)
-{
-    if (bWasSuccessful)
-    {
-        UE_LOG(LogTemp, Warning, TEXT("Response Code = %d"), Response->GetResponseCode());
-
-        if (!Response.IsValid())
-        {
-            UE_LOG(LogTemp, Warning, TEXT("Request failed!"));
-        }
-        else if (EHttpResponseCodes::IsOk(Response->GetResponseCode()))
-        {
-            TSharedPtr<FJsonObject> JsonObject = MakeShareable(new FJsonObject());
-            TSharedRef<TJsonReader<TCHAR>> JsonReader = TJsonReaderFactory<TCHAR>::Create(Response->GetContentAsString());
-            FJsonSerializer::Deserialize(JsonReader, JsonObject);
-
-            Token = JsonObject->GetStringField("token");
-
-            UE_LOG(LogTemp, Warning, TEXT("Token = %s"), *Token);
-            UploadFile("SaveGame1", 0);
-        }
-    }
-    else 
-    {
-        UE_LOG(LogTemp, Warning, TEXT("Request failed! Is the server up?"));
-    }
 }
 
 USaveGame* CloudySaveManagerImpl::Cloudy_LoadGameFromSlot(const FString& SlotName,
