+++ Source/CloudySaveManager/Private/CloudySaveManager.h
@@ -2,12 +2,8 @@
 #include "Kismet/BlueprintFunctionLibrary.h"
 #include "Engine.h"
 #include "GameFramework/SaveGame.h"
-#include "Http.h"
 
-//
-// Forward declarations.
-//
-//class USaveGame;
+DECLARE_LOG_CATEGORY_EXTERN(CloudySaveManagerLog, Log, All)
  
 class CloudySaveManagerImpl : public ICloudySaveManager
 {
@@ -18,26 +14,21 @@ public:
  
     /**
     *	Save the contents of the SaveGameObject to a slot, and upload it to CloudyGame's cloud.
-    *	@param SaveGameObject	Object that contains data about the save game that we want to write out
-    *	@param SlotName		Name of save game slot to save to.
-    *   @param UserIndex	For some platforms, master user index to identify the user doing the saving.
+    *	@param SaveGameObject   Object that contains data about the save game that we want to write out
+    *	@param SlotName		    Name of save game slot to save to.
+    *   @param UserIndex	    For some platforms, master user index to identify the user doing the saving.
     *   @param PCID             PlayerController object of the player to save
-    * 	@param IsAutosaved	Whether or not this save file is autosaved
-    *	@return			Whether we successfully saved this information
+    *
+    *	@return			        Whether we successfully saved this information
     */
     UFUNCTION(BlueprintCallable, Category = "CloudyGame")
-        bool Cloudy_SaveGameToSlot(USaveGame* SaveGameObject, const FString& SlotName, const int32 UserIndex, const int32 PCID); // APlayerController const* PC);
-
-    bool AttemptAuthentication(FString Username, FString Password);
-    bool UploadFile(FString Filename, int32 PlayerControllerId);
-    void OnResponseComplete(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bWasSuccessful);
-    void OnAuthResponseComplete(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bWasSuccessful);
+        bool Cloudy_SaveGameToSlot(USaveGame* SaveGameObject, const FString& SlotName, const int32 UserIndex, const int32 PCID);
 
     /**
-    *	Save the contents of the SaveGameObject to a slot.
-    *	@param SlotName			Name of save game slot to save to.
-    *   @param UserIndex		For some platforms, master user index to identify the user doing the saving.
-    * 	@param IsAutosaved		Whether or not the save file was autosaved
+    *	Download the save file from CloudyGame's cloud, then save the contents of the SaveGameObject to a slot.
+    *	@param SlotName			    Name of save game slot to save to.
+    *   @param UserIndex		    For some platforms, master user index to identify the user doing the saving.
+    *
     *	@return SaveGameObject		Object containing loaded game state (NULL if load fails)
     */
     UFUNCTION(BlueprintCallable, Category = "CloudyGame")
