+++ Source/CloudyWebAPI/Private/CloudyWebAPIPrivatePCH.h
@@ -1,0 +1,4 @@
+#include "ICloudyWebAPI.h"
+ 
+// You should place include statements to your module's private header files here.  You only need to
+// add includes for headers that are used in most of your module's source files though.
